+++ docs/CHANGES.md
@@ -45,6 +45,39 @@ All changes are logged here. These are auto-generated using the Merge & Complete
 
 
 
+
+
+
+
+
+
+
+
+
+
+
+## 2025-11-25 - Gavin Wallis - New Platform Feature
+- Add Integration Testing for the Core Module
+## 2025-11-25 - Test User - Service Fix
+- Finished
+## 2025-11-25 - Test User - New Service
+- Test entry for detail
+## 2025-11-25 - Test User - New Service
+- Test entry
+## 2025-11-25 - Test User - New Service
+- Finished all tasks
+## 2025-11-25 - Test User - New Service
+- Added test file for integration testing
+## 2025-11-25 - Test User - Service Fix
+- Finished
+## 2025-11-25 - Test User - New Service
+- Test entry for detail
+## 2025-11-25 - Test User - New Service
+- Test entry
+## 2025-11-25 - Test User - New Service
+- Finished all tasks
+## 2025-11-25 - Test User - New Service
+- Added test file for integration testing
 ## 2025-11-25 - Gavin Wallis - New Service
 - Add monitoring endpoint: nginx3 (network - 10.66.33.112)
 ## 2025-11-25 - Test User - Service Fix